requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
google-generativeai==0.8.3
python-multipart==0.0.12
jinja2==3.1.4
python-dotenv==1.0.1
Pillow==10.4.0